+++ openretina/cli/__main__.py
@@ -1,4 +1,4 @@
 if __name__ == "__main__":
-    from openretina.cli.run import main
+    from openretina.cli.openretina import main
 
     main()
