+++ openretina/utils/__init__.py
@@ -1,4 +1,4 @@
 from openretina.utils.file_utils import get_local_file_path
-from openretina.utils.plotting import save_stimulus_to_mp4_video
+from openretina.utils.plotting import plot_stimulus_composition, save_stimulus_to_mp4_video
 
-__all__ = ["get_local_file_path", "save_stimulus_to_mp4_video"]
+__all__ = ["get_local_file_path", "save_stimulus_to_mp4_video", "plot_stimulus_composition"]
