+++ pyproject.toml
@@ -33,7 +33,7 @@ dependencies = {file = ["requirements.txt"]}
 "openretina" = ["py.typed", "configs/*.yaml", "configs/**/*.yaml"]
 
 [project.scripts]
-openretina = "openretina.cli.run:main"
+openretina = "openretina.cli.openretina:main"
 
 [project.optional-dependencies]
 dev = [
